Refactor UserView component to include userProjectManager query key
