+++ app/(main)/loading.tsx
@@ -1,6 +1,6 @@
-import { LoadingElement } from "@/app/Components/Loading";
+import { LoadingElement, LoadingPage } from "@/app/Components/Loading";
 
 export default function Loading() {
     // You can add any UI inside Loading, including a Skeleton.
-    return <LoadingElement />;
+    return <LoadingPage />;
 }
