+++ package.json
@@ -34,6 +34,7 @@
         "@mui/x-data-grid-generator": "^7.18.0",
         "@mui/x-date-pickers": "^7.18.0",
         "@next/eslint-plugin-next": "^14.2.13",
+        "@parcel/watcher": "^2.4.1",
         "@react-email/components": "^0.0.25",
         "@react-email/render": "^1.0.1",
         "@tanstack/react-query": "^5.56.2",
